fix: new typeissues

+++ packages/dc_solver_pkg/tests/jax/test_aggregate_results.py
@@ -565,7 +565,7 @@ def test_aggregate_to_metric(mocker) -> None:
                     [3, 4, 5, 6, 7],
                     [10, 10, 10, 10, 10],
                 ],
-                dtype=float,
+                dtype=int,
             )
         ),
     )
@@ -946,7 +946,7 @@ def test_get_pst_switching_distance() -> None:
 
     # Case 2: PST optimization enabled but no initial taps provided
     optimized_taps = NodalInjOptimResults(
-        pst_tap_idx=jnp.array([[0, 1, 2, 3, 4]], dtype=float)  # shape: (n_timesteps=1, n_controllable_pst)
+        pst_tap_idx=jnp.array([[0, 1, 2, 3, 4]], dtype=int)  # shape: (n_timesteps=1, n_controllable_pst)
     )
     switching_distance = get_pst_switching_distance(optimized_taps=optimized_taps, initial_tap_idx=None)
     assert switching_distance == 0.0, "Switching distance should be 0 when initial tap indices are not provided"
@@ -954,7 +954,7 @@ def test_get_pst_switching_distance() -> None:
     # Case 3: No switching_distance - optimized taps match initial taps
     initial_tap_idx = jnp.array([2, 3, 4, 5, 6], dtype=int)
     optimized_taps = NodalInjOptimResults(
-        pst_tap_idx=jnp.array([[2, 3, 4, 5, 6]], dtype=float)  # shape: (n_timesteps=1, n_controllable_pst)
+        pst_tap_idx=jnp.array([[2, 3, 4, 5, 6]], dtype=int)  # shape: (n_timesteps=1, n_controllable_pst)
     )
     switching_distance = get_pst_switching_distance(optimized_taps=optimized_taps, initial_tap_idx=initial_tap_idx)
     assert switching_distance == 0.0, "Switching distance should be 0 when taps haven't changed"
@@ -962,7 +962,7 @@ def test_get_pst_switching_distance() -> None:
     # Case 4: Simple switching_distance case - single timestep
     initial_tap_idx = jnp.array([2, 3, 4, 5, 6], dtype=int)
     optimized_taps = NodalInjOptimResults(
-        pst_tap_idx=jnp.array([[3, 4, 5, 6, 7]], dtype=float)  # All shifted by +1
+        pst_tap_idx=jnp.array([[3, 4, 5, 6, 7]], dtype=int)  # All shifted by +1
     )
     switching_distance = get_pst_switching_distance(optimized_taps=optimized_taps, initial_tap_idx=initial_tap_idx)
     expected_switching_distance = 5.0  # Sum of (3-2)^2 + (4-3)^2 + (5-4)^2 + (6-5)^2 + (7-6)^2 = 1+1+1+1+1 = 5
@@ -973,7 +973,7 @@ def test_get_pst_switching_distance() -> None:
     # Case 5: Mixed positive and negative switching distances
     initial_tap_idx = jnp.array([5, 5, 5, 5, 5], dtype=int)
     optimized_taps = NodalInjOptimResults(
-        pst_tap_idx=jnp.array([[3, 7, 5, 4, 8]], dtype=float)  # switching distances: -2, +2, 0, -1, +3
+        pst_tap_idx=jnp.array([[3, 7, 5, 4, 8]], dtype=int)  # switching distances: -2, +2, 0, -1, +3
     )
     switching_distance = get_pst_switching_distance(optimized_taps=optimized_taps, initial_tap_idx=initial_tap_idx)
     expected_switching_distance = (
@@ -992,7 +992,7 @@ def test_get_pst_switching_distance() -> None:
                 [10, 10, 10, 10],  # Second timestep: 8^2 + 7^2 + 6^2 + 5^2 = 174
                 [0, 0, 0, 0],  # Third timestep: 2^2 + 3^2 + 4^2 + 5^2 = 54
             ],
-            dtype=float,
+            dtype=int,
         )  # shape: (n_timesteps=3, n_controllable_pst=4)
     )
     switching_distance = get_pst_switching_distance(optimized_taps=optimized_taps, initial_tap_idx=initial_tap_idx)
@@ -1007,7 +1007,7 @@ def test_get_pst_switching_distance() -> None:
         return get_pst_switching_distance(optimized_taps, initial_tap_idx)
 
     initial_tap_idx = jnp.array([1, 2, 3], dtype=int)
-    optimized_taps = NodalInjOptimResults(pst_tap_idx=jnp.array([[2, 3, 4]], dtype=float))
+    optimized_taps = NodalInjOptimResults(pst_tap_idx=jnp.array([[2, 3, 4]], dtype=int))
 
     switching_distance_jitted = jitted_switching_distance(optimized_taps, initial_tap_idx)
     switching_distance_normal = get_pst_switching_distance(optimized_taps, initial_tap_idx)
